requirements/runtime.txt
av
av==8.0.3; python_version < '3.7'
einops
facexlib
lmdb
mmcv-full>=1.3.13 # To support DCN on CPU
numpy
opencv-python!=4.5.5.62,!=4.5.5.64
# MMCV depends opencv-python instead of headless, thus we install opencv-python
# Due to a bug from upstream, we skip this two version
# https://github.com/opencv/opencv-python/issues/602
# https://github.com/opencv/opencv/issues/21366
# It seems to be fixed in https://github.com/opencv/opencv/pull/21382
Pillow
tensorboard
torch
torchvision
